refactor(mongo): centralize inline oplog replay
